Added SplatmapWater tool which renders water based on a splatmap
Similarly to heightmaps, the blue value is used for the height of the water at a given location (pixel), with a height_scale parameter to modify the scale. The alpha value is used for defining the transparency of the water at that pixel, with 0 meaning no water should be drawn at all.