basis for LOD implementation in test scenes
Scenes/Testing/TestWorld.tscn displays a first attempt at a LOD implementation - a lot is still hardcoded and not as efficient as it should be, it's just the groundwork.

Note: I am not pushing the test terrain heightmaps since they're very large, ask me for them if you want to run the scene!