Added a basic third person controller with mouse dragging camera
Can be activated by pressing '2' while in the Main scene and controlled by clicking and dragging.
Switch back to the flying first person controller by pressing '1'.
This third person camera currently acts as a Player, updating LOD -> TODO: Change this when the VR player is properly implemented and controls the LOD instead!