Added DynamicAssetHandler for syncing assets with server
The DynamicAssetHandler can keep track of the instances of a specific asset on the server. The local asset instances are kept in sync with this data.

A VestasWindmillHandler which implements the DynamicAssetHandler has been implemented. It keeps track of the asset with ID 1 (Vestas Windmills) and updates the local instances accordingly.