Moved ItemSpawner to own scene, used in 3rd person
The ItemSpawner has been moved to its own scene so that it can be reused. It is now used in the 3rd person PC perspective as well. For this, a 'MousePoint' has been added to it: This node is responsible for aligning the ray of the ItemSpawner with where the mouse is pointing.