Fixed regular assetpos requests
The request is now non-caching, since the result is expected to change regularly. Another issue which potentially stops requests if one was invalid has also been fixed.

Since the GRIDSIZE/2 problem was found not to actually be an issue (it was a miscalculation when testing), the corresponding code has been removed.