Increased cull margin on terrain for proper visibility
When the cull margin is too small, tiles disappear too early (since the renderer things they are not within the camera's view anymore). This extra cull margin should make them visible on the bottom edge on the screen as well.

There's a pull request in Godot for properly exposing custom mesh AABBs to the scripting language - once this is merged, we can actually create the proper AABB for that mesh instead of this somewhat unclean solution.