Fixed regression where ServerConnection was stuck if request failed
Since the json_cache was sometimes filled with a 'request in progress' flag which was never set to 'done', other threads with the same request could get stuck waiting forever. This was a problem with vegetation, where some requests are known to fail (the server wasn't filled with the according data yet).