Fixed bug where tooltip was displayed in main UI
The tooltip has a viewport which its sprite accesses - this viewport should only ever be visible through that sprite. However, previously, since the viewport was in a ViewportContainer, it was also an actual (tiny) UI element. This caused a small visual error and prevented some buttons to work, as it stopped the mouse click.