Workaround for flickering particles on some graphcis drivers
On my GTX 980 with driver version 430 on Ubuntu 19, the vegetation particles were flickering. This commit fixes it, although it relies on a bit of an ugly workaround:
Particle lifetimes can't be infinite, so it's currently limited to 1661992960 seconds. Realistically, this shouldn't ever be a problem, but it's not ideal.

The Godot issue for this problem is 18117 - once it's fixed, we should set the proper infinite lifetime as well.