Fixed delay in location tracker
The location tracker was changed to immediately get the position / make the screenshot as soon as it should. Only the actual request / saving the picture is done in a thread. That thread has medium priority to make sure it doesn't take too long.

In the process, the formatting of the script was updated according to our coding style.