Visual and QoL improvements in vegetation scaling
Vegetation modules now take a generic 1m-scaled billboard mesh and a
scale parameter, as opposed to having to manually create scaled
billboard meshes for each layer.

These scale parameters have been set to reflect the values in the
server. (We could add a request to get that scale parameter from the
server, so that we don't have to keep those synchronized manually, but
this works for now)

The particle shader now scales the billboard using the parameter.