Add AssetHandlerSpawner for dynamic loading
Asset handlers are now automatically created, which means that as long
as the asset has a scene at the right path, nothing needs to be done
manually for them to become visible!

The Assets singleton was expanded in the process, as well as some more
tweaks and debugging information which was helpful for testing.