Generalize basic asset handler functionality
A new AbstractAssetHandler was added, which DynamicAssetHandler and
LinearAssetHandler implement. This is because a lot of functionality
(periodically making server requests, keeping all assets in the response
instanced and deleting the rest, ...) is shared between them.

There is an issue with the streets: If they didn't get a good ground
position when they were instanced, they stay below the world since
they're never updated.