Workaround for initial streets below the ground
A delay can now be set for asset updates. Using that delay, the linear
asset handler now waits until the first update. After that wait time,
it's likely that there's a good obtainable ground position.

This should be revisited in the future since it's just a workaround that
could potentially fail.