Refactored input with _unhandled_input
All instances of _input were changed to _unhandled_input.

The ui controls were decoupled from the player controls since those
events are captured by the UI otherwise.

The mouse filter was changed in many Control elements to allow the mouse
to go through to the desired nodes.

A workaround was implemented in the ViewportContainers, it's still
unclear why it's necessary, but the input doesn't go to the player nodes
otherwise.