Add Perspective superclass with cleanup function
That cleanup function is e.g. used by the VR perspective to uninitialize
the ARVRInterface and to reset the viewport.

It doesn't fix all our problems with the ARVR viewports (see #87), but
it's still practical.

Note that the color space in VR is different now due to the
keep_3d_linear parameter - we will have to test what's better with the
actual VR headset.