Centralize all world shifting functionality
Instead of the multiple implementations we had for world shifting in
many different scripts (which were mostly identical), all world shifting
behavior is now centralized to handlers and groups.

There is still one issue with the player, which is why the
AbstractPlayer still has a custom world shifting implementation.

This will need a lot of testing to make sure everything works as good as
or better than before.