Add new icons to toolbar
NOTE: A ton of automatic Godot changes are also commited here. I think those make no actual difference.
If it does turn out this broke something on other machines, we can revert this commit and make the changes on a different machine.

Since the toolbar already uses the AutoTextureButton, its icons were updated.

The toolbar now has no ClientResources dependency. The same will gradually be done for the rest of the UI.