Add Singleton for Terrain
Since the terrain will always be requried in some form, it was added as a Singleton.

This way, it doesn't need to be instanced/removed at every state change, causing unnecessary loading.

The terrain starts loading as soon as the game was started to prevent trying to load invalid tiles at position 0, 0 at the start.