Improve detail texture (reduce tiling, blend with orthophoto)
This commit introduces some new concepts to the detail texture blending. I think it already looks better, but it's still a bit rough, there are a few things that could be improved (some have been mentioned in TODOs).

To reduce tiling and increase detail, a larger scale version of the detail texture is blended to the normal version when further away.

To make the detail texture look less like a carpet that's over the orthophoto, the hue and chroma of both are blended / taken into account for calculating how much the detail texture should show.