Add proof of concept for new road loader
This new node uses Geodot instead of the server.

I initially attempted modifying the LinearAssetHandler, but the differences in functionality turned out to be too big, so a completely new node was created. After testing this proof of concept, we'll see whether merging this with general AssetHandler stuff is feasible.