Generalize loaders + add loader for power poles
The AssetsFromShapefile loader class is now abstract, and it provides a function to implement which loads the fitting asset for a GeoPoint.

An implementation was added which looks at one attribute of the GeoPoint and can load different assets based on the value of this attribute in the individual GeoPoint.

This implementation is now used for loading power lines.