Add proof of concept new vegetation renderer
This renderer dynamically gets the heightmap for its position and thus works together with the terrain.

However, it does not actually load splatmap / phytocoenosis data yet - it's just a proof of concept for placing the plants.

The plant material will be extended to draw the correct plants in future commits.