Rebuild scenario/location; remove server dependency
Scenarios and their locations are now loaded from a shapefile referenced in geodata.json instead of the server.

Some game mode specific stuff was removed from the scenario in the process - thus, some functionality e.g. for the default wind direction and for the energy UI was commented out for now.