Add simple viewshed tool
Just a proof-of-concept - the basic functionality is given, but we'll want to add settings and make it more precise in the future.

The tool is usable by default. A simple symbol for it was added. When activated (eye symbol clicked), a 2km-radius viewshed analysis will be given at the position of the mouse cursor.