Move some Startup script stuff to project settings
I didn't set fullscreen to true in the settings because I think a window is actually more practical for testing. If we do want it to be (always) fullscreen, we can do that in the project settings.

The main reason for this change was that it was hard to figure out where settings came from - we want to have only one source (the project settings).