Apply random plant height (between min and max size)
A new approach is used for encoding the plant height: Previously, we scaled the billboard directly. This was static and lowered the resolution needlessly.

Now, we encode an individual plant's height (a random value between its min and max) within the distribution map. This way, the heights are varied, but that randomness is reproducable and even controllable if desired.

The distribution texture scale is also handled properly now, because it caused noticeable distortions since it didn't align with the plant grid.