Structural setup for new lake shader
Because the lake shading is closely related to terrain shading, it was added to the Realistic Terrain layer, not as a separate new layer type.

For this, the concept of "ExtraLODs" was added, which allow adding another mesh to shade a specific part of the terrain differently (while keeping the terrain itself there).