Display information about cursor on middle mouse click
Middle mouse click now opens a window. It shows the distance of the cursor from the camera (closes #267). In the future, more values could be added, e.g. regarding the Object under the mouse cursor.

In additional, the visuals of the cursor were made to be a bit more pleasant by not becoming huge nearby.