Initial (working) implementation of "teleport token"
Because of the special case that is the teleport token (it sends "new token", but really it means an updated player position), this broke some of the design, so there is some hacky code (marked with FIXME).
Overall though, it doesn't integrate too badly, just some minor restructurings are needed.