Optimize Buildings by reducing draw calls
This is a minor rework of how Buildings are generated: Shader parameters are used as little as possible; instead, the Vertex color attribute is set from the SurfaceTool to specify how the texture should be colored.

Loading times have been slightly improved as well by setting the material in _ready - loading nodes with Materials in threads seems to cause a stall (there are Godot issues for this).

Fairly noticeable difference on my machine.