Adapt and optimize vegetation with 4.0
This fixes the vegetation with 4.0 and makes some optimizations, some of which were overdue, and some which were made possible by 4.0. The biggest change is that the particle shader does as much work as possible rather than doing it in the draw shader later (which is called far more often).