fix: dolly camera transform was not registered correctly by position manager
As of commit ef4942989 the camera is positioned inside the scene tree and thus its position is not the same as global position.
Consequently, the position manager reads the wrong coordinates to render new LayerComposition features

This could be solved using the top level attribute on the camera and then pushing a position via a remote transform.

refs: ef4942989