feat: add AggregatedGameObjectAttribute and improve robustness of CalculatedGameObjectAttribute
the AggregatedGameObjectAttribute essentially allows defining an attribute whose value is read from a table: its value depends on the value of any number of other attributes.

the CalculatedGameObjectAttribute can now also set an attribute in the GeoFeature if needed, e.g. to visualize a value which is calculated by the game system (in our case, the amount of PV modules which fit in the given field, based on the set attributes).