feat: first implementation of "Bewirtschaftungsstreifen"
currently does not access real data but just uses a noise which is randomly rotated based on the LU ID.

Already quite a big improvement, the biggest limitation is that the stripe direction does not make sense.

Also, the effect doesn't make much sense on the plants themselves at the moment, since they are just darker. It'd probably be better to decrease their height or remove them, but we'd need better stripe data for that to look good.