feat: add far LOD to Terrain for much greater view distance
This new system is a generalization of the DetailLOD and the logic surrounding it.

Similarly to the DetailLOD, the FarLOD is a single plane. It's 40x40km and has a hole in the middle for the normal chunked system.

It reloads on its own, not threaded; some testing showed that it's fast enough to fit into our frame budget, if this is different on another device, the resolution could be changed.