fix: bugs and old leftovers in RepeatingObject height logic
The height was increasingly inaccurate the higher the distance between objects was.
It seems like it worked by chance for fences, but wasn't actually accurate.

There are 2 options now: get the height at the start of the instance (for e.g. traffic poles) and get the height at the middle of the instance (for e.g. fences - similar to how it was before).