feat: unify attribute-to-property definition
Instead of the tacked-on attributes_to_mesh_settings, there's now a way to define properties based on any attributes. A "priority" can be defined to decide which properties should override others.

This applies to LineObject and RepeatingObject.

It's fully backwards-compatible for RepeatingObject definitions, as the old way of configuring was kept (for now).