feat: new Road and Infrastructure rendering approach (LineObjectRenderer) & Height Overlay System (#355)
* feat: Streets are 3D objects which alter the terrain height

Largely working - very similar to the LID Overlay, which has been used successfully for a while.

Some remaining tasks to clean up left.

* feat: add HeightOverlay to RasterVegetation

* feat: adapt RoadLaneDirt to new height overlay

* feat/WIP: proof of concept for new Road renderer similar to RepeatingObject

* feat: add additional road segment objects

* refactor: abstract RepeatingRoags into LineObject

This new renderer shall be usable for anything where we want to instantiate a single object per line segment and stretch the object in the shader, as we now do for the new roads.

* refactor: line segment clean-up, shared Height Getter classes

* feat: various resource additions for new road systems

* fix: proper AABB handling of road segments

* fix: implement get_angle for ExactCurveHeightGetter

* feat: unify attribute-to-property definition

Instead of the tacked-on attributes_to_mesh_settings, there's now a way to define properties based on any attributes. A "priority" can be defined to decide which properties should override others.

This applies to LineObject and RepeatingObject.

It's fully backwards-compatible for RepeatingObject definitions, as the old way of configuring was kept (for now).

* refactor: clean up and document attribute->property settings

* refactor: replace multiline if with Dictionary

* fix: wrong word in LineObject description

* fix: make high overlay contribution pixels override lower ones by setting DEPTH

* refactor: rename LIDOverlay folder into Overlay

since it also contains HeightOverlay now

* refactor: naming and function abstraction for the Overlay systems

* refactor: add consistent and easily usable global signals for Height and LID Overlay