refactor: remove Terrain Decals
This was used for the overlay roads, but since they have been replaced by LineObjects, this is no longer needed.

Normal decals can be used for the terrain; this was a system that specifically rendered 3D objects top-down and projected them onto the terrain (At various resolutions for LOD). this was error-prone and not an intended use-case (gave out Warnings all the time).