Scattered Object Renderer (based on Raster Data) (#361)
* feat: first functional version of ScatteredObjectRenderer

first step for #358

This renderer can take either Scenes or Mesh resources and chooses its chunk accordingly.

Objects are placed if the value in a raster map match a certain expression, e.g. "8*" matches "8102".

Currently, only one object per match can be defined - this will have to be expanded.

Vector scatter layers are not yet supported.

The density is fixed; as noted in a comment, we could use blue noise with variable density to allow changing the density based on another raster map.

* feat: add "randomness" factor to ScatteredObjects

* feat: restructure config according to #358

* feat: implement Billboard system for mesh-based scattered objects

* feat: add wine Object and Billboard

* feat: generalize bird particles and add sparrow swarm

* refactor: address remarks from #361 review