perf: avoid reloading a feature multiple times within a single frame
If multiple `feature.changed` signals are emitted, then the connected functions are also called multiple times. Since we connect with `CONNECT_DEFERRED` here, this is useless: we only need to update the feature once, regardless of how many changes were made to it. This check uses the number of engine ticks to check whether a feature has already been added to the change queue this frame.